# Parquet iterator improvements